refactor ampliconv6 analysis flow into separate files. add helper for django setup / EMG_CONFIG fetching.
