+++ .talismanrc
@@ -35,6 +35,15 @@ fileignoreconfig:
 - filename: workflows/ena_utils/ena_auth.py
   checksum: 4920257f382f89084ed4a2ba3b88e7a5cff76ee23ddde9929ec6435e5b4c86fe
 
+- filename: workflows/flows/analyse_study_tasks/make_samplesheet_amplicon.py
+  allowed_patterns: [key]
+
+- filename: workflows/flows/analyse_study_tasks/set_post_analysies_states.py
+  allowed_patterns: [qc_passed_runs.csv]
+
+- filename: workflows/flows/analyse_study_tasks/sanity_check_amplicon_results.py
+  checksum: 6f0d0012faefa871ada95216a04997ce6ae54ff3e508b57399cccc859e1eccfa
+
 - filename: workflows/flows/assemble_study_tasks/assemble_samplesheets.py
   allowed_patterns: [key]
