+++ workflows/prefect_utils/datamovers.py
@@ -1,10 +1,9 @@
 from datetime import timedelta
 
-import django
 from prefect import flow
 from prefect.runtime import flow_run
 
-django.setup()
+import activate_django_first  # noqa
 
 from workflows.data_io_utils.filenames import file_path_shortener
 from workflows.prefect_utils.slurm_flow import EMG_CONFIG, run_cluster_job
